feat: :sparkles: Updated the brevo email sending class
An interface was created to be implemented when sending emails, necessary to reuse the code if it is necessary to use another provider. In the 'Brevo' email sending file, the input parameter validations were removed, as this is already done in the controller. Constants were also added for better organization of information, exception returns now received error codes for better identification of the response on the front-end.